% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstable.R
\name{get_descriptive_stats}
\alias{get_descriptive_stats}
\title{Calculate descriptive statistics for a data frame}
\usage{
get_descriptive_stats(data, group_by_cols, value_col, stats_settings)
}
\arguments{
\item{data}{Data frame that contains the data}

\item{group_by_cols}{Character object that holds the column names to be grouped}

\item{value_col}{Column name of the numeric data where the stats will be calculated on}

\item{stats_settings}{List of stats settings}
}
\value{
A list that contains the stats result requested with labels and depths for each group labels
}
\description{
This function takes a data frame and specifications of grouping and stats as input and
outputs a list which have two fixed components: label and depth, and several custom components which
are stores the stats specified in stats_settings.
}
